feat: implement complete authentication GUI with web components
- Create nerd-auth component with login/register forms
- Implement form submission to /auth endpoint
- Add mode toggling between login and registration
- Handle errors and auto-reload on success
- Update nerd-gui to conditionally render auth component
- Fix System.Authenticator registration in loadAuthenticator
- Fix Tag.ToMap() to convert NodeID to float64 for JSON compatibility
- Import nerd-auth only in nerd-gui for clean dependency structure

+++ api/msg/types.go
@@ -22,7 +22,7 @@ type Tag struct {
 // ToMap converts Tag to map[string]any for HTTP responses
 func (t *Tag) ToMap() map[string]any {
 	return map[string]any{
-		"nodeId": t.NodeID,
+		"nodeId": float64(t.NodeID),
 		"admin":  t.Admin,
 	}
 }
